fix: Wait a fraction of a second before adding files
Git was failing to add files because a lockfile was not deleted, thus causing issue #1.

This change should hopefully mitigate that problem in the future, but it's probably
flaky... I don't really want to bring the filesystem watcher into this, but I will if
I have to.

At any rate, at least there is some warning in the case of failure now.